Move LogoutButton to components folder
