+++ src/App.svelte
@@ -11,7 +11,7 @@
     import type { QueryCstNode } from './generated/chevrotain_dts';
     import Message from './Message.svelte';
     import SSOButton from './components/SSOButton.svelte';
-    import LogoutButton from './LogoutButton.svelte';
+    import LogoutButton from './components/LogoutButton.svelte';
     import SSORevokeButton from './SSORevokeButton.svelte';
     import { validate } from './session';
     import { parseDSL } from './dsl';
